requirements.txt
python-telegram-bot==21.9
aiomysql==0.2.0
python-dotenv==1.1.0
